Split architecture dependant and independant parts of FMA dot conversion
This PR splits FMA dot conversion from Triton GPU to LLVM in two parts:
- Common code with iteration across M/N dim
- Architecture dependant scalar multiplication of vectos across K dim
